Feature to search and report non- data member references
